Separate the implementation of REST api and Websocket Realtime API.
