+++ rocketchat-common/src/main/java/com/rocketchat/common/utils/Logger.java
@@ -1,0 +1,9 @@
+package com.rocketchat.common.utils;
+
+public interface Logger {
+    void info(String format, String... args);
+
+    void warning(String format, String... args);
+
+    void debug(String format, String... args);
+}
