+++ rocketchat-core/src/main/java/com/rocketchat/core/factory/ChatRoomFactory.java
@@ -1,7 +1,8 @@
 package com.rocketchat.core.factory;
 
 import com.rocketchat.common.data.model.BaseRoom;
-import com.rocketchat.core.RocketChatAPI;
+import com.rocketchat.core.ChatRoom;
+import com.rocketchat.core.RocketChatClient;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -13,16 +14,16 @@ import java.util.List;
 // TODO: 29/7/17 might have to make separate arraylist for each type, seems little impossible, better to keep generic
 public class ChatRoomFactory {
 
-    private RocketChatAPI api;
-    private List<RocketChatAPI.ChatRoom> rooms;
+    private RocketChatClient client;
+    private List<ChatRoom> rooms;
 
-    public ChatRoomFactory(RocketChatAPI api) {
-        this.api = api;
+    public ChatRoomFactory(RocketChatClient client) {
+        this.client = client;
         rooms = new ArrayList<>();
     }
 
-    private RocketChatAPI.ChatRoom createChatRoom(BaseRoom room) {
-        return api.new ChatRoom(room);
+    private ChatRoom createChatRoom(BaseRoom room) {
+        return new ChatRoom(client, room);
     }
 
     public ChatRoomFactory createChatRooms(List<? extends BaseRoom> roomObjects) {
@@ -35,18 +36,18 @@ public class ChatRoomFactory {
 
     public ChatRoomFactory addChatRoom(BaseRoom room) {
         if (getChatRoomByName(room.name()) == null) {
-            RocketChatAPI.ChatRoom newRoom = createChatRoom(room);
+            ChatRoom newRoom = createChatRoom(room);
             rooms.add(newRoom);
         }
         return this;
     }
 
-    public List<RocketChatAPI.ChatRoom> getChatRooms() {
+    public List<ChatRoom> getChatRooms() {
         return rooms;
     }
 
-    public RocketChatAPI.ChatRoom getChatRoomByName(String roomName) {
-        for (RocketChatAPI.ChatRoom room : rooms) {
+    public ChatRoom getChatRoomByName(String roomName) {
+        for (ChatRoom room : rooms) {
             if (room.getRoomData().name() != null
                     && roomName.contentEquals(room.getRoomData().name())) {
                 return room;
@@ -55,8 +56,8 @@ public class ChatRoomFactory {
         return null;
     }
 
-    public RocketChatAPI.ChatRoom getChatRoomById(String roomId) {
-        for (RocketChatAPI.ChatRoom room : rooms) {
+    public ChatRoom getChatRoomById(String roomId) {
+        for (ChatRoom room : rooms) {
             if (roomId.contentEquals(room.getRoomData().roomId())) {
                 return room;
             }
@@ -65,7 +66,7 @@ public class ChatRoomFactory {
     }
 
     public Boolean removeChatRoomByName(String roomName) {
-        for (RocketChatAPI.ChatRoom room : rooms) {
+        for (ChatRoom room : rooms) {
             if (room.getRoomData().name() != null
                     && roomName.contentEquals(room.getRoomData().name())) {
                 return rooms.remove(room);
@@ -75,7 +76,7 @@ public class ChatRoomFactory {
     }
 
     public Boolean removeChatRoomById(String roomId) {
-        for (RocketChatAPI.ChatRoom room : rooms) {
+        for (ChatRoom room : rooms) {
             if (room.getRoomData().roomId().equals(roomId)) {
                 return rooms.remove(room);
             }
@@ -83,7 +84,7 @@ public class ChatRoomFactory {
         return false;
     }
 
-    public Boolean removeChatRoom(RocketChatAPI.ChatRoom room) {
+    public Boolean removeChatRoom(ChatRoom room) {
         return rooms.remove(room);
     }
