+++ rocketchat-core/src/test/java/com/rocketchat/core/TestUtils.java
@@ -12,7 +12,7 @@ import io.fabric8.mockwebserver.dsl.TimesOnceableOrHttpHeaderable;
 
 public class TestUtils {
     @SafeVarargs
-    public static void setupMockServer(RocketChatAPI api, DefaultMockServer server,
+    public static void setupMockServer(RocketChatClient api, DefaultMockServer server,
                                        Pair<Object, String>... expectEmit) throws InterruptedException {
         final CountDownLatch latch = new CountDownLatch(1);
         final boolean[] connected = {false};
